Move utils methods into base package
