+++ src/server/storage_node.rs
@@ -8,9 +8,9 @@ use tokio_util::codec::length_delimited;
 
 use log::{debug, error};
 
+use crate::base::utils::node_id_from_address;
 use crate::generated::get_root_as_generic_request;
 use crate::server::router::request_router;
-use crate::utils::node_id_from_address;
 use std::net::SocketAddr;
 use std::path::Path;
 use std::sync::Arc;
