+++ src/storage/raft_group_manager.rs
@@ -1,9 +1,9 @@
+use crate::base::message_utils::accessed_inode;
+use crate::base::utils::{finalize_request_without_prefix, LengthPrefixedVec};
 use crate::generated::*;
-use crate::message_utils::accessed_inode;
 use crate::peer_client::PeerClient;
 use crate::server::LocalContext;
 use crate::storage::raft_node::{node_contains_raft_group, RaftNode};
-use crate::utils::{finalize_request_without_prefix, LengthPrefixedVec};
 use flatbuffers::FlatBufferBuilder;
 use futures_util::future::FutureExt;
 use std::collections::HashMap;
